Added Foreign keys to Meetup model class

+++ DevMeetups/ViewModels/MeetupsFormViewModel.cs
@@ -9,7 +9,7 @@ namespace DevMeetups.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Category { get; set; }
+        public byte Category { get; set; }
         public IEnumerable<Category> Categories { get; set;}
     }
 }
